v5/v5/internal api/requirements.txt
Flask>=3.0.0
requests>=2.31.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1
gunicorn>=22.0.0
PyJWT>=2.8.0
